requirements.txt
Flask
google-cloud-texttospeech
google-generativeai
Werkzeug
gunicorn